services/ai_service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "minio>=7.2.20",
    "numpy>=2.4.1",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "uvicorn>=0.40.0",
]
